# docs/api/OPENAPI.yaml
openapi: 3.1.0
info:
  title: Caladdin API
  version: 1.0.0
  description: |
    HTTP API for Caladdin — AI scheduling assistant.

    **Route groups covered:** `/api/*`, `/book/*`, `/s/*`, `/jobs/*`

    **Authentication:**
    - **Session cookie** (`caladdin_session`) — host-facing routes under `/api/*`
    - **API key** (`x-api-key` header) — cron/worker routes under `/jobs/*`
    - **Guest action token** (signed HMAC, query or body) — cancel/reschedule on `/s/:token/*`
    - **Public** — `/book/*` metadata and `/s/:token` HTML pages (no auth)

    Base URL is `CALADDIN_BASE_URL` (e.g. `https://caladdin.onrender.com`).
servers:
  - url: https://caladdin.onrender.com
    description: Production (example)
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: caladdin_session
      description: HMAC-signed session cookie set by `GET /auth/callback`
    apiKey:
      type: apiKey
      in: header
      name: x-api-key
      description: Value of `CALADDIN_API_KEY` env var
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    UserProfile:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        email:
          type: string
          format: email
        timezone:
          type: string
          example: America/Chicago
        privacyMode:
          type: string
          enum: [private, trusted, open]
        workingHoursStart:
          type: string
          pattern: '^\d{2}:\d{2}$'
          example: '09:00'
        workingHoursEnd:
          type: string
          pattern: '^\d{2}:\d{2}$'
          example: '18:00'
        onboardingComplete:
          type: boolean
        calendarConnected:
          type: boolean
    ProfilePatch:
      type: object
      properties:
        timezone:
          type: string
          minLength: 1
          maxLength: 64
        privacyMode:
          type: string
          enum: [private, trusted, open]
        workingHoursStart:
          type: string
          pattern: '^\d{2}:\d{2}$'
        workingHoursEnd:
          type: string
          pattern: '^\d{2}:\d{2}$'
      description: At least one field required
...
    SchedulingSession:
      type: object
      properties:
        id:
          type: string
          format: uuid
        token:
          type: string
        host_user_id:
          type: string
          format: uuid
        slots:
          type: array
          items:
            type: object
            properties:
              start:
                type: string
                format: date-time
              end:
                type: string
                format: date-time
              score:
                type: number
        host_name:
          type: string
          nullable: true
        context:
          type: string
          nullable: true
        posture:
          type: string
        status:
          type: string
        proposed_event_ids:
          type: array
          items:
            type: string
        expires_at:
          type: string
          format: date-time
        invitee_email:
          type: string
          nullable: true
        host_timezone:
          type: string
          nullable: true
        duration_minutes:
          type: integer
          nullable: true
        offered_slots:
          type: array
          nullable: true
        selected_slot:
          type: object
          nullable: true
        google_event_id:
          type: string
          nullable: true
        proposed_alternatives:
          type: array
    PublicBookingHost:
      type: object
      properties:
        name:
          type: string
        username:
          type: string
        timezone:
          type: string
    PublicBookingEventType:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        durationMinutes:
          type: integer
        description:
          type: string
          nullable: true
        availabilityRules:
          type: object
          additionalProperties: true
        schedulingMode:
          type: string
          enum: [single, round_robin]
...
paths:
  /api/profile:
    get:
      tags: [Profile]
      summary: Get host profile
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Profile view
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags: [Profile]
      summary: Update host profile (onboarding, timezone, hours)
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfilePatch'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          description: Invalid or empty payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Not authenticated
        '500':
          description: Save failed
  /api/sessions:
    get:
      tags: [Sessions]
      summary: List scheduling sessions for authenticated host
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Session list
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: '#/components/schemas/SchedulingSession'
        '401':
          description: Not authenticated
...
  /api/calendar.ics:
    get:
      tags: [Calendar]
      summary: Host ICS feed of confirmed bookings
      description: Returns `text/calendar` attachment for authenticated host sessions.
      security:
        - sessionCookie: []
      responses:
        '200':
          description: ICS calendar file
          content:
            text/calendar:
              schema:
                type: string
        '401':
          description: Not authenticated
  /book/{username}/{slug}:
    parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
        description: Host username (lowercase)
      - name: slug
        in: path
        required: true
        schema:
          type: string
        description: Event type slug (lowercase)
    get:
      tags: [Public Booking]
      summary: Public booking page metadata
      description: |
        Returns host and event-type metadata for the public booking SPA (`web/booking.js`).
        No authentication required.
      responses:
        '200':
          description: Booking page data
          content:
            application/json:
              schema:
                type: object
                properties:
                  host:
                    $ref: '#/components/schemas/PublicBookingHost'
                  eventType:
                    $ref: '#/components/schemas/PublicBookingEventType'
        '404':
          description: Booking page not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /book/{username}/{slug}/slots:
    get:
      tags: [Public Booking]
      summary: Available booking slots for event type
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: daysAhead
          in: query
          schema:
            type: integer
            default: 14
            maximum: 60
      responses:
        '200':
          description: Slot list with assigned host for round-robin types
        '404':
          description: Booking page not found
...